to do: add lux filter

+++ sensor_device/main/device.h
@@ -1,4 +1,5 @@
 #include "drivers/i2c/i2c.h"
+#include "drivers/light_sensor/light_sensor.h"
 #ifndef  HEADER_DEVICE_H
 #define  HEADER_DEVICE_H
 
@@ -6,13 +7,12 @@
 #include <freertos/FreeRTOS.h>
 #include <freertos/task.h>
 #include <stdint.h>
-#include <stdint.h>
 
 struct data {
-    double  temperature,
-            humidity,
-            pressure,
-            lux;
+    uint32_t temperature,
+             humidity,
+             pressure,
+             lux;
 };
 
 #endif
